move base_maps setter into shared module
