+++ ecoscope/platform/tasks/results/_pydeck_map.py
@@ -11,11 +11,7 @@ from pydantic.json_schema import SkipJsonSchema
 from wt_registry import register
 
 from ecoscope.platform.annotations import AdvancedField, AnyGeoDataFrame
-from ecoscope.platform.tasks.results._map_utils import (
-    DEFAULT_TILE_LAYER_PRESETS,
-    TileLayer,
-    make_preset_or_custom_json_schema_extra,
-)
+from ecoscope.platform.tasks.results._map_utils import TileLayer
 
 PYDECK_CUSTOM_LIBRARIES = [
     {
@@ -326,29 +322,6 @@ class BitmapLayerDefinition(BaseModel):
     bounds: list[float]
     opacity: float = 1.0
     legend: LegendSegment | SkipJsonSchema[None] = None
-
-
-_preset_or_custom_json_schema_extra = make_preset_or_custom_json_schema_extra(TileLayer, DEFAULT_TILE_LAYER_PRESETS)
-
-
-@register()
-def set_base_maps_pydeck(
-    base_maps: Annotated[
-        list[TileLayer] | SkipJsonSchema[None],
-        Field(
-            json_schema_extra=_preset_or_custom_json_schema_extra,
-            title=" ",
-            description="Select tile layers to use as base layers in map outputs. \
-            The first layer in the list will be the bottommost layer displayed.",
-        ),
-    ] = None,
-) -> Annotated[list[TileLayer], Field()]:
-    if base_maps is None:
-        base_maps = [
-            TileLayer(layer_name="TERRAIN"),
-            TileLayer(layer_name="SATELLITE", opacity=0.5),
-        ]
-    return base_maps
 
 
 @dataclass
